refactor builtin/index
